Moving methods from Preprocessor to OASParsers.
